fix clippy

+++ crates/rpc/rpc-eth-api/src/helpers/fee.rs
@@ -162,7 +162,7 @@ pub trait EthFees: LoadFee {
 
 
                 for header in &headers {
-                    base_fee_per_gas.push(header.base_fee_per_gas.unwrap_or_default() as u128);
+                    base_fee_per_gas.push(header.base_fee_per_gas.unwrap_or_default());
                     gas_used_ratio.push(header.gas_used as f64 / header.gas_limit as f64);
                     base_fee_per_blob_gas.push(header.blob_fee().unwrap_or_default());
                     blob_gas_used_ratio.push(
@@ -201,9 +201,9 @@ pub trait EthFees: LoadFee {
                         .chain_spec()
                         .base_fee_params_at_timestamp(last_header.timestamp)
                         .next_block_base_fee(
-                            last_header.gas_used as u128,
-                            last_header.gas_limit as u128,
-                            last_header.base_fee_per_gas.unwrap_or_default() as u128,
+                            last_header.gas_used ,
+                            last_header.gas_limit,
+                            last_header.base_fee_per_gas.unwrap_or_default() ,
                         ),
                 );
